[CI] Do not export test matrix. (#21191)
The pwsh used on windows is based on dotnet framework. Dotnet framework
has a limit in the size of the env that can be passed to a child
process. We are very close to reach the limit, but after #21009 we can
ignore that env variable since it is parsed from the args.

This is just needed for the ci pipeline because the issues are going to
have are related with the vsdrops task used to upload nuget results
etc..