[msbuild] Add @(MaciOSXcodeProject) action
Introduces a `@(MaciOSXcodeProject)` build action which can be used
to build and consume the outputs of Xcode framework projects.

The following metadata is supported on this item:

    <MaciOSXcodeProject Include="path/to/project.xcodeproj">
      <Configuration>Release</Configuration>
      <Kind>Framework</Kind>
      <OutputPath></OutputPath>
      <ReferenceXcFramework>true</ReferenceXcFramework>
      <SchemeName></SchemeName>
      <SharpieBind>true</SharpieBind>
      <SharpieNamespace></SharpieNamespace>
      <SmartLink>true</SmartLink>
    </MaciOSXcodeProject>

 * `%(Configuration)`: The name of the configuration to use to build
    the project. The default value is `Release`.

 * `%(OutputPath)`: Can be set to override the XCARCHIVE and XCFRAMEWORK
    output path of the Xcode project.

 * `%(ReferenceXcFramework)`: Output XCFRAMEWORK files will be added as
    a `@(NativeReference)` to be bound or otherwise used by the project.
    Metadata supported by `@(NativeReference)` like `%(Kind)` or
	`%(SmartLink)` will be forwarded if set.

 * `%(SchemeName)`: The name of the build scheme that should be used to
    build the project. The default value is
    `$(IntermediateOutputPath)xcode/{SchemeName}-{Hash}`.

 * `%(SharpieBind)`: Output XCFRAMEWORK files will be bound using
    Objective Sharpie. The default value is `true`.

 * `%(SharpieNamespace)`: The namespace to use when binding output
    XCFRAMEWORK files with Objective Sharpie.

A new `_BuildMaciOSXcodeProjectFrameworks` target will attempt to build
XCARCHIVE and XCFRAMEWORK files for each `@(MaciOSXcodeProject)` item.
These outputs will be created for the platform specified in the target
framework. If multiple target frameworks are specified, the project
will be built for each platform during each inner build.

The built XCFRAMEWORK files will be added as `@(NativeReference)` items
automatically, unless the `%(ReferenceXcFramework)` metadata is set to
`false`.

A new `$(MaciOSPrepareForBuildDependsOn)` build extension point has been
added to allow customer projects to more easily hook into the beginning
of the build process.