[ScriptingBridge] Remove outdated tests and add documentation. Fixes #21242.
The test we had in our codebase for the SBApplication class was not only
outdated, but also incorrect.

With recent changes the test code doesn't work anymore, so delete it.

Instead, add XML documentation that explains how to use the class in a way
that works.

Fixes https://github.com/xamarin/xamarin-macios/issues/21242.