[CI] Remove the APIScan stage from the main ci/pr pipeline.
We are trying to reduce the stages so that tests are triggered to just
once. Due to a limitation from VSTS we have to reduce the number of
stages in the main ci, which is the one that triggers the tests.

We already have created two new pipelines that will run the API
scan as soon as packages are built therefore we do not longer need
the stage to be ran.