[Rgen] Provide a data model to make the code generator more performant.
In oder to fully take advange of the
[caching](https://github.com/dotnet/roslyn/blob/main/docs/features/incremental-generators.md#caching) provided by the
incremental code generator infrastructure we move to use a data model
that will represent the code changes that will trigger a regeneration of
certain classes.

The current implementation only takes care of the changes that will
trigger a regeneration on enums:

1. Enum fully qualified name changed.
2. Attributes added or removed in the enum.
2. Enum values added/removed.

We provide our own Comparer to have more control on how the structures
are compared.