[dotnet] Fix generating the list of supported target platform versions.
Fix how we compute the target platforms we support:

1. Start with the list of target frameworks we support ("net8.0-ios17.0 net8.0-ios18.0 net9.0-ios19.0")
2. Split on the space.
3. Remove the "netX.Y-" part.
4. Remove duplicates.

Now we're left with the list if target platform versions we support.

The bug was in step 3: we only removed netX.Y where X.Y was the current .NET version (currently 9.0), and not previous .NET versions.

This meant that for the sample input from above, we'd be left with this incorrect list:

* net8.0-ios17.0
* net8.0-ios18.0
* ios18.0