[msbuild] Fix some issues with various types of resources when building from Windows. (#21586)
While adding a test for a new feature, it turned out that the new test uncovered existing bugs. So here I'm adding that new test, and fixing those bugs.

The main part is that computing the virtual project path for resource items on Windows wasn't correct, in particular when referencing resources outside the current project directory. The existing code didn't even have enough information to compute the correct values...

Sidenote: the virtual project path is used to decide where in an app bundle a particular resource should go.

So:

* Add new metadata to resource items, for the path of the defining project + the path of the main project. This is required to compute the correct virtual project path.
* Completely rewrite the code to compute virtual project path, prioritizing making the code legible and easy to understand (in particular minimizing differences in code paths between macOS and Windows), and add lots of comments.
* Add unit tests for computing the virtual project path.
* Add a new .NET project test for all variations of resources we support.
* Run said new test on Windows as well, both in remote and local (Hot Restart) mode.