[xtro] Don't remove fields we've found from existing list of managed fields.
The same native field may be declared in multiple frameworks, and if we remove
it from the list of managed fields we've matched, then we'll removed the
binding as missing when we find it in the second framework.

Instead keep track of the fields we've matched in a separate hash table, and
use that to figure out which fields haven't been matched.