[msbuild] Validate entitlements in CompileEntitlements.
A very common problem customers run into is when they build for release, and
don't update the entitlements their app requests correctly.

In particular this happens often with the `aps-environment` entitlement, which
must be "development" for local development, and "production" when publishing
to the App Store. Fortunately the error the App Store is fairly actionable,
but the reverse isn't true: using "production" for local development will just
fail with a generic error about not being able to install the app on the
device.

So add validation for the entitlements. We currently only validate the
`aps-environment` entitlement, but this can easily be expanded to more
entitlements in the future.

Additionally we now log all the entitlements a provisioning profile grants,
which will help with debugging customer problems.