[bgen] Add support for [return: Release] for types subclassing NativeObject or implementing INativeObject. (#21752)
Add support for [return: Release] for types subclassing NativeObject or implementing INativeObject.

This was implemented by always passing a value for the "owns" parameter when
creating a managed wrapper for a given native pointer (and the passed in value
for the "owns" parameter honors the presence of the Release attribute)

This also required adding an overload taking an "owns" parameter for a few
types which didn't already have such an overload.