[CI] Do not double checkout macios-adr to keep the crendetials.
Doing a second checkout is error prone. One of the reason is that there
is no guarantee that any of our other scripts is not going to remove a
random temp directory (actually, the API diff make does remove it).

If the checkout directory is removed, the post checkout action from VSTS
will fail with something of the kind:

```
[error]An error occurred trying to start process '/usr/local/bin/git'
with working directory
'/Users/builder/azdo/_work/22/a/change-detection/tmp/src/macios-adr'. No
such file or directory
```

The correct way to workaround the issue is to 'mimic' the way VSTS uses
git via https.extraheaders (note that this might not work with old
versions of git lfs, but should work with versions newer than 2.7.1).

We add a script that will update the configuration of a vsts git repo to
use an extra base auth header with the system.token of the pipeline,
which should have access to the repo.

I have added some extra docs in the older script for github so that
other developers understand the usage and will not make the same mistake
with API diff again.