[CI] Do not double checkout macios-adr to keep the credentials. (#21807)
Doing a second checkout is error prone. One of the reason is that there
is no guarantee that any of our other scripts is not going to remove a
random temp directory (actually, the API diff make does remove it).

If the checkout directory is removed, the post checkout action from VSTS
will fail with something of the kind:

```
[error]An error occurred trying to start process '/usr/local/bin/git'
with working directory
'/Users/builder/azdo/_work/22/a/change-detection/tmp/src/macios-adr'. No
such file or directory
```

The reset script was doing the right thing by setting the http extra headers but it had a typo in the variable name.