[tests] Don't run 'make' in 'tests/tests-libraries' as part of test setup.
It's not necessary; this already happens on the bots before the test run
anyway, so any work done here is redundant on the bots. It's even timing out
(after two minutes) on some test runs, which means this should save up to two
minutes for the msbuild tests on the bots.