[RGen] Add the platform availability as part of the property data model.
Add the data in the model as well as the availability data to the
accessors. We have not modified the events data model in this PR, we
only focused on properties.

Added ToString implementations to simplify the test debugger and fixed a
bug in one of the constructors in which the structure was not creating a
copy of the dictionary but was keeping the ref that was passed to the
structure, this resulted in undesirable side effects when modifying the
collection.