[RGen] Add the field data to the enum values structure. (#21856)
Since we are calculating which enum values were added as part of the
code change we can retrieve at that time the field data information.
This way we reduce the number of times we query the semantic model in
the code generation. This opens the door to use the code changes
structure to generate the smart enums and reduce the number of allocated
objects.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>