[Rgen] Provide a way to parse the generic BindingTypeAttribute and a union struct.
We want to be able to store as parte of the code change struct any
information related to the binding type. To do so we have a new
BindingTypeAttribute that is generic to allow extra config.

We also provide a new struct that works as a union of all the supported
generic BindingTypeAttribute, that way we can add that new struct (in a
follow up PR) to the code change struct.