[registrar] Fix registration of stub Objective-C classes. Fixes #21546. (#21593)
Objective-C has a fairly new feature called stub (unrealized) classes, which don't exist until runtime. They were introduced to support Swift classes exposed to Objective-C, when the exact layout of the class isn't known at compile time, and the creation of the class itself is deferred until runtime.

An Objective-C class is marked as a stub class by adding the `objc_class_stub` attribute to the class definition:

```objc
__attribute__((objc_class_stub)
@interface ObjCStubClass : NSObject
@end
```

Sidenote: the class will also have the `objc_subclassing_restricted` attribute, because Clang requires the latter when the former is present.

When this attribute is present, Clang will generate code to load the class at runtime, instead of just fetching it, when doing `[ObjCStubClass class]`.

Our immediate symptom of stub classes when using the static registrar is that the above would crash:

* We don't have the original headers for any third-party Objective-C code, so we generate it based on the information we have available.
* We don't add the `objc_class_stub` attribute to the generated Objective-C header (because we don't know that the class is a stub class).
* Clang compiles the wrong code for `[ObjCStubClass class]`, because of the missing attribute.

The first step in the fix is to add a new `IsStubClass` property to the `Register` attribute, so that we can let the registrars know about them.

Then update the static registrar to add the `objc_class_stub` attribute to the generated Objective-C header.

We also need a fix for the dynamic registrar, because we call the `objc_getClass ("ObjCStubClass")` there to get the Objective-C class, and that doesn't work unless the stub class has already been realized/used (`objc_getClass` returns NULL). It turns out that computing the total number of Objective-C classes in the process will also realize all stub Objective-C classes, so that's what we do when we find a stub Objective-C class.

References:

* https://github.com/swiftlang/swift/blob/main/docs/ObjCInterop.md#stub-classes
* https://alwaysprocessing.blog/2023/02/16/objc-unrealized-classes

Fixes https://github.com/xamarin/xamarin-macios/issues/21546.