[Rgen] Do not create an emitter per code change. (#21883)
Emitter should not store internal state that will changes based on the
processed code changes. This means we can reduce the number of objects
we create to the minimum amount. This change will reduce the memory
usage specially when building the Apple SDK which has 100s of classes to
generated.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>