[ObjCRuntime] Ensure all assemblies' module constructors have executed before failing registrar lookup. (#21885)
The managed static registrar depends on each assembly registering itself to
find types from that assembly. This is done using a module constructor, but
the module constructor isn't necessarily executed before the managed static
registrar needs the assembly registered (in particular for app extensions,
which don't have a managed entry point).

That would lead to exceptions like this:

    *** Terminating app due to uncaught exception 'ObjCRuntime.RuntimeException', reason: 'Could not find the type 'ObjCRuntime.__Registrar__' in the assembly '...'. (ObjCRuntime.RuntimeException)
		at ObjCRuntime.RegistrarHelper.GetMapEntry(String assemblyName)
		at ObjCRuntime.RegistrarHelper.GetMapEntry(Assembly assembly)
		[...]

Fix this by manually calling the module constructor for every loaded assembly
if we can't find it in the map of registered assemblies.