[CI] Batch developer builds in the CI to use resources better.
This is annoying because there is a diff between PR and CI trigger. PRs
will cancel previous builds (which was giving problems due to the auto
format). However, now we have the other problem around in which we are
executing the build several times. We cannot cancel, but we can batch
the builds.

The last commit will always be build.