[CI] Allow the commits to be written in PRs that were started with a CI trigger
Now that we allow PRs to be executed for our team members via a ci
trigger the script that writtes the comment does so in the commit (as we
do in CI). This breaks the way we are used to work. To fix this problem
the github script has been fixed to write the comment in the prs
associated with a commit if they are open.