[CI] Allow the comments to be written in PRs that were started with a CI trigger (#21921)
Now that we allow PRs to be executed for our team members via a ci
trigger the script that writes the comment does so in the commit (as we
do in CI). This breaks the way we are used to work. To fix this problem
the github script has been fixed to write the comment in the prs
associated with a commit if they are open.