[CI] Improve the IsPR method to account for other trigger types.
The PR CI might get executed due to other triggers that are not the CI
one, for example the commnad '/azp run' uses a PR trigger and not a CI
one.

The IsPR method had a bug in which it would correctly return true due to
the trigger used for the build, but the PRIds array would be empty.

If we find a trigger that represents a PR we now update the PR Ids
accordingly. This fixes an error in which we would get the following:

```pwsh
InvalidOperation: Cannot index into a null array.
```

An example of the error can be found in the build:
https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=10815634&view=logs&j=600d49c6-e96b-5e8b-2c3a-9f351e2e5848&t=f7d6b16e-fe1e-5522-9638-9daa7906620a