[RGen] Do not pass the root binding context around. (#21945)
There is no need to pass the root binding around to be able to generate
the Library.g.cs. What we can do is change the code to have to diff
incremental code provides.

1. A provider that will detect code changes.
2. A provider that will detect changes that required a Library.g.cs
change.

To achieve the above what we do is to add the library information when
we detect code changes. This new fields that provide the library
information are then collected to trigger a new generation event.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>
Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>