[Rgen] Add missing test data for tvOS and MacCatalyst. (#21955)
xUnit will fail test theories that have no data. To fix these failures
we add data for all platforms in those tests that was missing it.

The only time now in which we will have an issue with the theories will
be when no platforms are built, which makes no sense.