[RGen] Use the correct attribute for field properties. (#21959)
We cannot use Export because we need to be ablet to pass the library
name to generate the property field correctly. This change moves from
using Export<Field> to Field<Property> which allows to mark a property
as a field AND will allow use to add the library name.

This change + the one in PR
https://github.com/xamarin/xamarin-macios/pull/21945 improves the way
the Library.g.cs file is generated.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>