[xtro] Skip native enum values that aren't available on the current platform. (#21958)
This also required a fix in cecil-tests: don't skip obsolete members when
verifying availability attributes.

We end up failing a test that verifies that a member must exist on a platform,
if it has an availability attribute claiming so in other platforms' assemblies
- the problem being that the test wouldn't find the member on a platform it
was merely obsoleted (but still existing!) there.