[RGen] Keep track of the SpecialType and the MetadataName in the TypeInfo struct. (#21976)
This change allows use to detect during the code generation if the user
used the nint or nuint types or instead used IntPtr and UIntPtr. Later
we can use that information to perform the correct native calls when
needed.

It also improves the implementation of IsVoid since we simply have to
check its SpecialType.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>