[Rgen] Add a biding syntax factory that will generated C# expressions.
Add the initial implementation of a syntaxt factory that can be used to
generate the C# calls needed by the bindings.

This implementation only contains the Getter methods needed to retrieve
fields in order to keep PRs in a reasonable size.

Example usage when writting a bindings:

```csharp
sb.AppendRaw($@"
if ({property.Name} is null)
    {property.Name} = {FieldConstantGetter (property)};
return {property.Name}
";
```

The API takes advantes of the fact that the data model has all the
needed infor and the makes the code cleaner when importing it with a
statuc using.