[Rgen] Add a binding syntax factory that will generate C# expressions. (#21980)
Add the initial implementation of a syntax factory that can be used to
generate the C# calls needed by the bindings.

This implementation only contains the Getter methods needed to retrieve
fields in order to keep PRs in a reasonable size.

Example usage when writing bindings:

```csharp
sb.AppendRaw($@"
if ({property.Name} is null)
    {property.Name} = {FieldConstantGetter (property)};
return {property.Name}
";
```

The API takes advantage of the fact that the data model has all the
needed info and the makes the code cleaner when importing it with a
status using.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>
Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>