[Rgen] Add support to the syntax factory to generate setters for fields. (#21983)
This commit has the following important changes to carefully review:

1. We move to make the method name otpional in the GetConstant and
SetConstant. This is done by the CallerMemberName attribute which will
add the name automatically to match the caller. This is done at compile
time and is equal to nameof. This way we reduce typing and typos.
2. Added a tuple to return the getter and setter. This way we make sure
that when we add support for a new kind of field, that we add both.
3. For those fields that we cannot set, we return a throw expression.
Ideally this will never happen because we will make sure the analyzer
stops us from trying to use Set properties on unsupported type.
4. Added support for casting in the Setter. We are using a lambda
factory that captures the type of the property ONLY when is needed, this
should be lazy enough.

With this change we should be able to fully generate field properties.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>