[Rgen] Renamed CodeChanges to Binding to be more clear. (#22031)
Initially the structure was name to make it simpler to understand that
it was a code change/update that was detected by the roslyn incremental
generator. Now that the structure is shared with the transformer, it
makes more sense to rename it to 'Binding'.

We have update all the helper classes too.