[Rgen] Use a roslyn code generator to help generate the flag parsing of all attributes.
The old Xamarin API used A LOT of attributes as flags to indicate
different generating methods. All this attributes do not have any data
and can be used as a feature flag.

For example, if we wanted to generated a static class we can do:
```csharp

[Static]
interface MyStaticClass {}
```
In this commit we are relaying in a custom roslyn code generator that
parses an attribute and generates the flag properties for our data
model. All the code is very repetitive and specially error prone. The
generator allows use just to focus on the name of the attribute and the
targets to which the attribute can be applied.

When you do the review you'll notice that the targets used in the
AttributeNames in the transformer do not match the ones in the
bgen/Attributes.cs file. The mismatch is because people back in the day
were careless and used the default flag, which is All. That is a
mistake, so I went throgh the trouble of using the correct target and
add a document string explaining the use of the flag and therefore the
reasoning of the updated target.

A sample of the generated code can be found
[here](https://gist.github.com/mandel-macaque/ed4277457a3afa6606af0f78ee9bf07e)