[RGen] Update the transformer generator to access the attr data.
Add new code that will generate a 2 methods for each BindingAttribute
marked AttributeName. A method that will allow to check if the attribute data is
present and a method that allows to retrieve the data.

The new generated code will look liek this https://gist.github.com/mandel-macaque/7ffb4b215cb5856d91fd391393564170