[RGen] Add the code generation for the data attributes. (#22053)
This change adds two more properties to the data models when they are
compile. One to check if the data attr is present, and a second one with
the attribute details.

The boolean property is decorated acordingly to help with the
nullability analisys.

An example of the generated code can be found here:
https://gist.github.com/mandel-macaque/04185ea481ca30ef9ff2ecf6dbf51ea0

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>