[Rgen] Add factory method that creates a NSArray aux variable.
To be able to send arrays to objc we need to convert them to a NSArray.
This change adds a factory method that will create the aux variables for
when we need that conversion.

The factory allows to prepend the declaration with a using caluse to
simplify the memory management in the bindings.