[Rgen] Move all the attributes to be readonly record struct.
There is a reason we landed the code first as a struct, later as a
readonly record struct. In the first pass, we added the manual code, had
full control and were sure the method did what we exepcted. In the
second pass we remove all the manual code EXCEPT that code that should
be manual. We knew what to change becuase the tests failed with the code
generated by the compiler.

The only method that the compiler generates that is not quite correct is
the Equals for the BaseTypeAttribute that has to be written by hand.

Tests verify that the behaviour is the same as that one of the manual
code.