[Rgen] Add properties in the Property data model to decide if a field is needed.
Add some new properties, to be only used by the generator, in the
property data model to decide the inclusion of a backing field and if
this field needs to be checked.

We have moved the ToString method to the .Generator and .Transformer
files so that we can reflect the two new properties. The same has
happened for the Equals method.