[Rgen] Add the needed code to parse a BindAsAttriubte.
This change is a little more interesting than the others because I had
to update the way we retrieve the attributes. Return attrbutes are only
present on a special collection for IMethodSymbols, this meant that we
had to add an extra pass for those attributes in the generic method in
pur extensions.